1) Add cpplint.py which checks for C++ style according to google's C++ style guide. 2) Added memory_mapped_file and file_info which will be used by the blob_manager which is the next task. 3) Had to bump up the boost version to 1.56.0 because there is memory mapped file bug in previous version of boost that prevents the compilation on VS 2013. The link of the bug is https://svn.boost.org/trac/boost/ticket/9369. Plus it is the latest version and will have more features and bug fixes. 4) Fixed compiler warnings and cpplint.py warning in some of the files.

+++ src/jonoondb_api/index_info.cc
@@ -17,7 +17,7 @@ struct IndexInfo::IndexInfoData {
 };
 
 IndexInfo::IndexInfo(const char* name, int16_t type, const char* columns[],
-                     int columnsLength, bool isAscending) {
+                     size_t columnsLength, bool isAscending) {
   m_indexInfoData = new IndexInfoData();
   if (name != nullptr) {
     m_indexInfoData->Name = name;
